Fixed rounding issue with rot. angle, fixed angle input box layout

+++ backend/src/main/java/nl/wilcokas/luckystackworker/ij/histogram/LswImageMetadata.java
@@ -27,5 +27,5 @@ public class LswImageMetadata {
     private int cropHeight;
     private int zoomFactor;
     private ChannelEnum channel;
-    private int angle;
+    private double angle;
 }
